Semana11Sesion1/dlopez/requirements.txt
backpack==0.1
blinker==1.4
cleo==0.6.8
click==8.0.1
colorama==0.4.4
Faker==0.8.18
Flask==2.0.1
flask-orator==0.2.0
inflection==0.3.1
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
orator==0.9.9
pastel==0.1.1
pendulum==1.5.1
psycopg2==2.9.1
pyaml==16.12.2
Pygments==2.9.0
pylev==1.4.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
simplejson==3.17.3
six==1.16.0
text-unidecode==1.2
tzlocal==1.5.1
Werkzeug==2.0.1
wrapt==1.12.1
